reaftc: melhorando o codigo: adicionando serviço para as rotas que redenriza dados do banco
