+++ src/controllers/BlogController.php
@@ -88,31 +88,6 @@ class BlogController extends PostsModel
       ]);
    }
 
-   static public function viewPost($id)
-   {
-      $result = PostsModel::findPostsById($id);
-
-      if (isset($result['error']) && $result['error']) {
-         return [
-            "error" => true,
-            "data" => $result,
-            "message" => "Erro ao buscar dados"
-        ];
-      }
-
-      if (isset($result) && sizeof($result) === 0) {
-         return [
-            "error" => false,
-            "data" => $result,
-            "message" => "Não foi encontrado dados"
-        ];
-      }
-      return [
-         "error" => false,
-         "data" => $result[0]
-      ];
-   }
-
    static public function getPostByUser($id)
    {
       $result = PostsModel::findPostsByUser($id);
@@ -137,32 +112,6 @@ class BlogController extends PostsModel
          "error" => false,
          "data" => $result
      ]);
-   }
-
-   static public function viewPostByUser($id)
-   {
-      $result = PostsModel::findPostsByUser($id);
-
-      if (isset($result['error']) && $result['error']) {
-         return [
-            "error" => true,
-            "data" => $result,
-            "message" => "Erro ao buscar dados"
-        ];
-      }
-
-      if (sizeof($result) === 0) {
-         return [
-            "error" => true,
-            "data" => $result,
-            "message" => "Não foi encontrado dados"
-        ];
-      }
-      
-      return [
-         "error" => false,
-         "data" => $result
-      ];
    }
 
    static public function insertPost()
